requirements.txt
celery==5.5.3
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.14.0
psycopg2-binary==2.9.10
redis==6.2.0
